enhancement: separate logic into hook
